pyproject.toml
[project]
name = "llm-forwarder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "asyncio>=3.4.3",
    "dask[complete,distributed]>=2025.5.1",
    "distributed>=2025.5.1",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydub>=0.25.1",
    "pygetwindow>=0.0.9",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "uvicorn>=0.34.2",
    "websockets>=15.0.1",
]

[tool.ruff]
target-version = "py312"
line-length = 240
indent-width = 4

# Enable linters
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "N",  # pep8-naming
    "SIM", # flake8-simplify
    "T20", # flake8-print
    "PL", # pylint
]

# Ignore specific rules
ignore = ["D100"]  # 禁用缺少模块文档字符串的检查

# Allow autofix for all enabled rules (when `--fix`) is provided
fixable = ["ALL"]
unfixable = []

# Exclude files and directories
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.isort]
known-third-party = ["fastapi", "pydantic", "loguru", "dask", "uvicorn"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false

[tool.pylint.messages_control]
disable = [
    "C0114",  # missing-module-docstring
    "C0115",  # missing-class-docstring
    "C0116",  # missing-function-docstring
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.11.12",
]
